Point configs to new data
* update several hardcodes
* refactor the harcodes to limit their re-use
