+++ external/fv3fit/fv3fit/emulation/data/transforms.py
@@ -22,6 +22,40 @@ TensorDataset = Mapping[Hashable, tf.Tensor]
 AnyDataset = Mapping[Hashable, NumericContainer]
 
 
+def _register_derived_variables(mapping: DerivedMapping, timestep):
+    # Watch out for these hardcoded names below. these should be inferred from
+    # the data or configuration.
+    @mapping.register(
+        "tendency_of_air_temperature_due_to_microphysics",
+        required_inputs=["air_temperature_input", "air_temperature_output"],
+    )
+    def _(self):
+        end = self._mapper["air_temperature_output"]
+        begin = self._mapper["air_temperature_input"]
+        return (end - begin) / timestep
+
+    @mapping.register(
+        "tendency_of_specific_humidity_due_to_microphysics",
+        required_inputs=["specific_humidity_input", "specific_humidity_output"],
+    )
+    def tendency_of_specific_humidity_due_to_microphysics(self):
+        end = self._mapper["specific_humidity_output"]
+        begin = self._mapper["specific_humidity_input"]
+        return (end - begin) / timestep
+
+    @mapping.register(
+        "tendency_of_cloud_water_mixing_ratio_due_to_microphysics",
+        required_inputs=[
+            "cloud_water_mixing_ratio_input",
+            "cloud_water_mixing_ratio_output",
+        ],
+    )
+    def tendency_of_cloud_water_mixing_ratio_due_to_microphysics(self):
+        end = self._mapper["cloud_water_mixing_ratio_output"]
+        begin = self._mapper["cloud_water_mixing_ratio_input"]
+        return (end - begin) / timestep
+
+
 def open_netcdf_dataset(path: str) -> xr.Dataset:
     """Open a netcdf from a local/remote path"""
 
@@ -36,7 +70,8 @@ def derived_dataset(
     all_variables: Sequence[str], dataset: AnyDataset, tendency_timestep_sec: int = 900
 ):
 
-    derived = DerivedMapping(dataset, microphys_timestep_sec=tendency_timestep_sec)
+    derived = DerivedMapping(dataset)
+    _register_derived_variables(derived, tendency_timestep_sec)
     dataset = derived.dataset(all_variables)
 
     return dataset
