+++ external/vcm/vcm/derived_mapping.py
@@ -16,13 +16,8 @@ class DerivedMapping(Mapping):
     VARIABLES: MutableMapping[Hashable, Callable[..., xr.DataArray]] = {}
     REQUIRED_INPUTS: MutableMapping[Hashable, Iterable[Hashable]] = {}
 
-    def __init__(
-        self, mapper: Mapping[Hashable, xr.DataArray], microphys_timestep_sec: int = 900
-    ):
+    def __init__(self, mapper: Mapping[Hashable, xr.DataArray]):
         self._mapper = mapper
-        # used for derived tendency variables in the microphysics
-        # TODO: should this just be part of the post-processing?
-        self._microphys_timestep_sec = microphys_timestep_sec
 
     @classmethod
     def register(cls, name: Hashable, required_inputs: Iterable[Hashable] = None):
@@ -253,39 +248,6 @@ def pQ2(self):
         return self._mapper["pQ2"]
     except KeyError:
         return xr.zeros_like(self["pressure_thickness_of_atmospheric_layer"])
-
-
-@DerivedMapping.register(
-    "tendency_of_air_temperature_due_to_microphysics",
-    required_inputs=["air_temperature_input", "air_temperature_output"],
-)
-def tendency_of_air_temperature_due_to_microphysics(self):
-    end = self._mapper["air_temperature_output"]
-    begin = self._mapper["air_temperature_input"]
-    return (end - begin) / self._microphys_timestep_sec
-
-
-@DerivedMapping.register(
-    "tendency_of_specific_humidity_due_to_microphysics",
-    required_inputs=["specific_humidity_input", "specific_humidity_output"],
-)
-def tendency_of_specific_humidity_due_to_microphysics(self):
-    end = self._mapper["specific_humidity_output"]
-    begin = self._mapper["specific_humidity_input"]
-    return (end - begin) / self._microphys_timestep_sec
-
-
-@DerivedMapping.register(
-    "tendency_of_cloud_water_mixing_ratio_due_to_microphysics",
-    required_inputs=[
-        "cloud_water_mixing_ratio_input",
-        "cloud_water_mixing_ratio_output",
-    ],
-)
-def tendency_of_cloud_water_mixing_ratio_due_to_microphysics(self):
-    end = self._mapper["cloud_water_mixing_ratio_output"]
-    begin = self._mapper["cloud_water_mixing_ratio_input"]
-    return (end - begin) / self._microphys_timestep_sec
 
 
 @DerivedMapping.register("internal_energy", required_inputs=["air_temperature"])
