+++ projects/microphysics/train/all-tendency-limited.yaml
@@ -14,14 +14,14 @@ loss:
       learning_rate: 0.0001
     name: Adam
   loss_variables:
-  - air_temperature_output
-  - specific_humidity_output
-  - cloud_water_mixing_ratio_output
+  - air_temperature_after_precpd
+  - specific_humidity_after_precpd
+  - cloud_water_mixing_ratio_after_precpd
   - total_precipitation
   weights:
-    air_temperature_output: 500000.0
-    specific_humidity_output: 500000.0
-    cloud_water_mixing_ratio_output: 1.0
+    air_temperature_after_precpd: 500000.0
+    specific_humidity_after_precpd: 500000.0
+    cloud_water_mixing_ratio_after_precpd: 1.0
     total_precipitation: .04
 model:
   architecture:
@@ -35,13 +35,9 @@ model:
   direct_out_variables:
   - total_precipitation
   residual_out_variables:
-    air_temperature_output: air_temperature_input
-    specific_humidity_output: specific_humidity_input
-    cloud_water_mixing_ratio_output: cloud_water_mixing_ratio_input
-  tendency_outputs:
-    air_temperature_output: tendency_of_air_temperature_due_to_microphysics
-    specific_humidity_output: tendency_of_specific_humidity_due_to_microphysics
-    cloud_water_mixing_ratio_output: tendency_of_cloud_water_mixing_ratio_due_to_microphysics
+    air_temperature_after_precpd: air_temperature_input
+    specific_humidity_after_precpd: specific_humidity_input
+    cloud_water_mixing_ratio_after_precpd: cloud_water_mixing_ratio_input
   normalize_key: mean_std
   enforce_positive: true
   selection_map:
